pyproject.toml
[tool.poetry]
packages = [
    { include = "functions" },
    { include = "services" },
    { include = "util" },
]
name = "micro_cam"
version = "0.1.0"
description = "Auto Cam Processing Tool"
authors = [
    "Josh S Wilkinson <JoshsWilkinson@Outlook.com>",
]
license = "Proprietary"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
numpy = "^1.24.2"
numpy-stl = "^3.0.0"
fastapi = "^0.95.0"
pyopencl = "^2022.3"
uvicorn = { extras = [
    "standard",
], version = "^0.21.1" }
pandas = "^1.5.2"
python-multipart = "^0.0.5"
httpx = "^0.23.3"
contourpy = "^1.0.7"
cython = "3.0.0b1"
setuptools = "^67.4.0"
files-com = "^1.0.925"


[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
ipykernel = "^6.21.2"
scipy = "^1.10.1"
matplotlib = "^3.7.1"

[build-system]
requires = [
    "poetry-core",
]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "main:server.run"
compute = "util.debugger:run_compute_test"
clear = "pwsh:Clear-Temp"
cy_build = "cy_build:run_setup"            # not working yet
# cy build |= python .\cy_build.py build_ext --inplace
